- all design arch modules in shared
