+++ bisqapps/androidClient/build.gradle.kts
@@ -70,7 +70,8 @@ android {
 }
 
 dependencies {
-//    implementation(project(":shared:domain"))
+    implementation(project(":shared:presentation"))
+    implementation(project(":shared:domain"))
     implementation(project(":shared:ui"))
     debugImplementation(compose.uiTooling)
 }
