+++ bisqapps/androidNode/build.gradle.kts
@@ -52,7 +52,8 @@ android {
 }
 
 dependencies {
-    //    implementation(project(":shared:domain"))
+    implementation(project(":shared:presentation"))
+    implementation(project(":shared:domain"))
     implementation(project(":shared:ui"))
     debugImplementation(compose.uiTooling)
 }
