+++ bisqapps/androidNode/src/androidMain/kotlin/network/bisq/mobile/android/node/MainActivity.kt
@@ -6,6 +6,8 @@ import androidx.activity.compose.setContent
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.tooling.preview.Preview
 import network.bisq.mobile.*
+import network.bisq.mobile.domain.data.repository.GreetingRepository
+import network.bisq.mobile.presentation.MainPresenter
 
 class MainActivity : ComponentActivity() {
     // TODO use a DI framework to provide implementations
