+++ bisqapps/iosClient/iosClient/ContentView.swift
@@ -1,5 +1,7 @@
 import UIKit
 import SwiftUI
+import presentation
+import domain
 import ui
 
 struct ComposeView: UIViewControllerRepresentable {
