+++ bisqapps/shared/domain/src/androidMain/kotlin/network/bisq/mobile/domain/Platform.android.kt
@@ -1,0 +1,9 @@
+package network.bisq.mobile.domain
+
+import android.os.Build
+
+class AndroidPlatform : Platform {
+    override val name: String = "Android ${Build.VERSION.SDK_INT}"
+}
+
+actual fun getPlatform(): Platform = AndroidPlatform()
