+++ bisqapps/shared/domain/src/androidUnitTest/kotlin/network/bisq/mobile/domain/Test.android.kt
@@ -1,0 +1,12 @@
+package network.bisq.mobile.domain
+
+import kotlin.test.Test
+import kotlin.test.assertTrue
+
+class AndroidGreetingTest {
+
+    @Test
+    fun testExample() {
+        assertTrue(Greeting().greet().contains("Android"), "Check iOS is mentioned")
+    }
+}
