--- bisqapps/shared/domain/src/commonMain/kotlin/Dummy.kt
@@ -1,4 +1,0 @@
-// To avoid build failing in iOS
-interface Dummy {
-    val name: String
-}
