+++ bisqapps/shared/domain/src/commonMain/kotlin/network/bisq/mobile/domain/Greeting.kt
@@ -1,0 +1,17 @@
+package network.bisq.mobile.domain
+
+open class Greeting {
+    protected val platform = getPlatform()
+
+    open fun greet(): String {
+        return "Hello, ${platform.name}!"
+    }
+}
+
+interface GreetingFactory {
+    fun createGreeting(): Greeting
+}
+
+class DefaultGreetingFactory : GreetingFactory {
+    override fun createGreeting() = Greeting()
+}
