+++ bisqapps/shared/domain/src/commonMain/kotlin/network/bisq/mobile/domain/Platform.kt
@@ -1,0 +1,7 @@
+package network.bisq.mobile.domain
+
+interface Platform {
+    val name: String
+}
+
+expect fun getPlatform(): Platform
