+++ bisqapps/shared/domain/src/commonTest/kotlin/network/bisq/mobile/domain/Test.kt
@@ -1,0 +1,12 @@
+package network.bisq.mobile.domain
+
+import kotlin.test.Test
+import kotlin.test.assertTrue
+
+class CommonGreetingTest {
+
+    @Test
+    fun testExample() {
+        assertTrue(Greeting().greet().contains("Hello"), "Check 'Hello' is mentioned")
+    }
+}
