+++ bisqapps/shared/domain/src/iosMain/kotlin/network/bisq/mobile/domain/Platform.ios.kt
@@ -1,0 +1,9 @@
+package network.bisq.mobile.domain
+
+import platform.UIKit.UIDevice
+
+class IOSPlatform: Platform {
+    override val name: String = UIDevice.currentDevice.systemName() + " " + UIDevice.currentDevice.systemVersion
+}
+
+actual fun getPlatform(): Platform = IOSPlatform()
