+++ bisqapps/shared/domain/src/iosTest/kotlin/network/bisq/mobile/domain/Test.ios.kt
@@ -1,0 +1,12 @@
+package network.bisq.mobile.domain
+
+import kotlin.test.Test
+import kotlin.test.assertTrue
+
+class IosGreetingTest {
+
+    @Test
+    fun testExample() {
+        assertTrue(Greeting().greet().contains("iOS"), "Check iOS is mentioned")
+    }
+}
