+++ bisqapps/shared/presentation/build.gradle.kts
@@ -1,10 +1,7 @@
-
 plugins {
     alias(libs.plugins.kotlinMultiplatform)
     alias(libs.plugins.kotlinCocoapods)
     alias(libs.plugins.androidLibrary)
-    alias(libs.plugins.jetbrainsCompose)
-    alias(libs.plugins.compose.compiler)
 }
 
 kotlin {
@@ -20,27 +17,25 @@ kotlin {
     iosSimulatorArm64()
 
     cocoapods {
-        summary = "Shared UI for all bisqapps"
-        homepage = "Link to the Shared Module homepage"
+        summary = "Shared Presentation Logic, navigation and connection between data and UI"
+        homepage = "X"
         version = "0.0.1"
         ios.deploymentTarget = "16.0"
         podfile = project.file("../../iosClient/Podfile")
         framework {
-            baseName = "ui"
+            baseName = "presentation"
             isStatic = false
         }
     }
 
     sourceSets {
         commonMain.dependencies {
-            implementation(compose.runtime)
-            implementation(compose.foundation)
-            implementation(compose.material)
-            implementation(compose.ui)
-            implementation(compose.components.resources)
-            implementation(compose.components.uiToolingPreview)
-            implementation(libs.androidx.lifecycle.viewmodel)
-            implementation(libs.androidx.lifecycle.runtime.compose)
+            //put your multiplatform dependencies here
+            implementation(project(":shared:domain"))
+            implementation(project(":shared:ui"))
+
+            implementation(libs.logging.kermit)
+            implementation("org.jetbrains.kotlinx:kotlinx-coroutines-core:1.9.0")
         }
         commonTest.dependencies {
             implementation(libs.kotlin.test)
