--- bisqapps/shared/ui/src/androidMain/kotlin/network/bisq/mobile/Platform.android.kt
@@ -1,9 +1,0 @@
-package network.bisq.mobile
-
-import android.os.Build
-
-class AndroidPlatform : Platform {
-    override val name: String = "Android ${Build.VERSION.SDK_INT}"
-}
-
-actual fun getPlatform(): Platform = AndroidPlatform()
