--- bisqapps/shared/ui/src/commonMain/kotlin/network/bisq/mobile/Greeting.kt
@@ -1,17 +1,0 @@
-package network.bisq.mobile
-
-open class Greeting {
-    protected val platform = getPlatform()
-
-    open fun greet(): String {
-        return "Hello, ${platform.name}!"
-    }
-}
-
-interface GreetingFactory {
-    fun createGreeting(): Greeting
-}
-
-class DefaultGreetingFactory : GreetingFactory {
-    override fun createGreeting() = Greeting()
-}
