--- bisqapps/shared/ui/src/commonMain/kotlin/network/bisq/mobile/MainPresenter.kt
@@ -1,30 +1,0 @@
-package network.bisq.mobile
-
-import kotlinx.coroutines.flow.MutableStateFlow
-import kotlinx.coroutines.flow.StateFlow
-
-/**
- * Main Presenter as an example of implementation for now.
- */
-class MainPresenter(private val greetingRepository: GreetingRepository) : BasePresenter(), AppPresenter {
-    // Observable state
-    private val _isContentVisible = MutableStateFlow(false)
-    override val isContentVisible: StateFlow<Boolean> = _isContentVisible
-
-    private val _greetingText = MutableStateFlow("Welcome!")
-    override val greetingText: StateFlow<String> = _greetingText
-
-    fun refresh() {
-        _greetingText.value = greetingRepository.getValue()
-    }
-
-    // Toggle action
-    override fun toggleContentVisibility() {
-        _isContentVisible.value = !_isContentVisible.value
-    }
-
-    override fun onResume() {
-        super.onResume()
-        refresh()
-    }
-}
