--- bisqapps/shared/ui/src/commonMain/kotlin/network/bisq/mobile/Platform.kt
@@ -1,7 +1,0 @@
-package network.bisq.mobile
-
-interface Platform {
-    val name: String
-}
-
-expect fun getPlatform(): Platform
