--- bisqapps/shared/ui/src/commonTest/kotlin/network/bisq/mobile/Test.kt
@@ -1,12 +1,0 @@
-package network.bisq.mobile
-
-import kotlin.test.Test
-import kotlin.test.assertTrue
-
-class CommonGreetingTest {
-
-    @Test
-    fun testExample() {
-        assertTrue(Greeting().greet().contains("Hello"), "Check 'Hello' is mentioned")
-    }
-}
