--- bisqapps/shared/ui/src/iosMain/kotlin/network/bisq/mobile/Platform.ios.kt
@@ -1,9 +1,0 @@
-package network.bisq.mobile
-
-import platform.UIKit.UIDevice
-
-class IOSPlatform: Platform {
-    override val name: String = UIDevice.currentDevice.systemName() + " " + UIDevice.currentDevice.systemVersion
-}
-
-actual fun getPlatform(): Platform = IOSPlatform()
