--- bisqapps/shared/ui/src/iosTest/kotlin/network/bisq/mobile/Test.ios.kt
@@ -1,12 +1,0 @@
-package network.bisq.mobile
-
-import kotlin.test.Test
-import kotlin.test.assertTrue
-
-class IosGreetingTest {
-
-    @Test
-    fun testExample() {
-        assertTrue(Greeting().greet().contains("iOS"), "Check iOS is mentioned")
-    }
-}
